Move PlanPrinter to sql.planner.planPrinter package
extraction of inner-classes as package-private top-level
classes will follow.
