+++ presto-main/src/test/java/com/facebook/presto/sql/planner/assertions/PlanAssert.java
@@ -17,7 +17,7 @@ import com.facebook.presto.Session;
 import com.facebook.presto.metadata.Metadata;
 import com.facebook.presto.sql.planner.Plan;
 
-import static com.facebook.presto.sql.planner.PlanPrinter.textLogicalPlan;
+import static com.facebook.presto.sql.planner.planPrinter.PlanPrinter.textLogicalPlan;
 import static java.lang.String.format;
 
 public final class PlanAssert
